% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{active_cases_dataset}
\alias{active_cases_dataset}
\title{Bulgaria active cases dataset}
\format{
An object of class \code{list} of length 1.
}
\usage{
active_cases_dataset
}
\description{
Bulgaria active cases dataset
}
\keyword{datasets}
