% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{bulgaria_grid}
\alias{bulgaria_grid}
\title{Custom Bulgaria grid used by geofacet}
\format{
An object of class \code{data.frame} with 27 rows and 4 columns.
}
\usage{
bulgaria_grid
}
\description{
Custom Bulgaria grid used by geofacet
}
\keyword{datasets}
