% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{common_chart_theme}
\alias{common_chart_theme}
\title{Constructs a theme with common configurations used for the project charts.}
\usage{
common_chart_theme()
}
\value{
ggplot2 theme
}
\description{
Constructs a theme with common configurations used for the project charts.
}
