% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_date_cases_dataframe}
\alias{create_date_cases_dataframe}
\title{Constucts a date cases dataframe}
\usage{
create_date_cases_dataframe(dataset)
}
\value{
data frame
}
\description{
Constucts a date cases dataframe
}
