% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_date_percentage_dataframe}
\alias{create_date_percentage_dataframe}
\title{Constructs a date percentage data frame}
\usage{
create_date_percentage_dataframe(dataset)
}
\value{
data frame
}
\description{
Constructs a date percentage data frame
}
