% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_cases.R
\name{create_disease_chronology_chart}
\alias{create_disease_chronology_chart}
\title{Creates a disease chronology chart}
\usage{
create_disease_chronology_chart()
}
\description{
Creates a disease chronology chart
}
