% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week_cases.R
\name{create_week_places_cases_chart}
\alias{create_week_places_cases_chart}
\title{Creates week places cases chart.}
\usage{
create_week_places_cases_chart(per100k = FALSE, facet = FALSE)
}
\arguments{
\item{per100k}{Whether to use per 100000 cases values}

\item{facet}{Whether to facet the chart}
}
\description{
Creates week places cases chart.
}
