% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{date_cases_dataset}
\alias{date_cases_dataset}
\title{Bulgaria date cases dataset}
\format{
An object of class \code{list} of length 9.
}
\usage{
date_cases_dataset
}
\description{
Bulgaria date cases dataset
}
\keyword{datasets}
