% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{export}
\alias{export}
\title{Exports a plot as an image.}
\usage{
export(plot, name, ext = ".svg", path = "out/", width = 16, height = 9)
}
\arguments{
\item{plot}{The plot to be exported}

\item{name}{Name of the exported image file}

\item{ext}{Image file extension}

\item{path}{Path to the storage directory}

\item{width}{Width of the exported image}

\item{height}{Height of the exported image}
}
\description{
Exports a plot as an image.
}
