% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{latest_totals_dataset}
\alias{latest_totals_dataset}
\title{Bulgaria latest cases}
\format{
An object of class \code{list} of length 15.
}
\usage{
latest_totals_dataset
}
\description{
Bulgaria latest cases
}
\keyword{datasets}
