% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{positive_tests_dataset}
\alias{positive_tests_dataset}
\title{Bulgaria positive tests dataset}
\format{
An object of class \code{list} of length 281.
}
\usage{
positive_tests_dataset
}
\description{
Bulgaria positive tests dataset
}
\keyword{datasets}
