% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{week_places_cases_dataset}
\alias{week_places_cases_dataset}
\title{Bulgaria week places cases dataset}
\format{
An object of class \code{data.frame} with 1080 rows and 8 columns.
}
\usage{
week_places_cases_dataset
}
\description{
Bulgaria week places cases dataset
}
\keyword{datasets}
